backend/pyproject.toml
[tool.poetry]
name = "keypoint-backend"
version = "0.1.0"
description = "Flask BFF of keypoint detection sample app"
authors = ["Sony Semiconductor Solutions Corporation"]
license = "Apache-2.0"
packages = [ { include = "src" }, ]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.2"
flask-cors = "^3.0.10"
flatbuffers = "1.11.0"
Flask-SocketIO = "^5.3.5"
python-dotenv = "^1.0.0"
console-access-api = { path = "./lib/console-access-api/", develop = false }
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-mock = "^3.11.1"
pytest-cov = "^4.1.0"
Sphinx = "^7.1.2"
black = "^23.7.0"
pyproject-flake8 = "^6.0.0.post1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 99
force-exclude = '/SmartCamera'

[tool.flake8]
max-line-length = 99
exclude = [".venv", ".git", "__pycache__", "**/aitrios_console.py", "SmartCamera"]
max-complexity = 10
ignore = "E203,E266,W503,E402"